Minor Update
Changes to floatmath,
heap,
sdefs,
fpu,
graphics pixel bounds fixed

+++ src/hmem/smem/heap.h
@@ -1,4 +1,5 @@
 #pragma once
-
+#include <stddef.h>
+#include <stdint.h>
 void heap_init(unsigned long long memsize);
-void* kalloc( int bytes );
+void* kalloc( uint64_t bytes );
